Move modal to chat component
