+++ frontend/chat-plugin/src/airyRenderProps/AiryHeaderBar/index.module.scss
@@ -78,20 +78,3 @@
   background: var(--color-light-gray);
 }
 
-.buttonWrapper {
-  height: 24px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  button {
-    margin: 0 8px;
-  }
-}
-
-  .cancelButton{
-  
-  }
-  
-  .endChatButton{
-  
-  }
